optimize dataset loading and embedding parsing - Fix loadText to properly split corpus into sentences by line boundaries - Replace per-word encodeTokens calls with direct map lookup via encodeSingleToken - Switch string keys from std::string to std::string_view to eliminate heap allocations - Replace strtod with fast_float::from_chars for faster float parsing in loadEmbeddingsToMatrix - Fix trainOnCorpus to take sentences by value instead of const ref to allow move semantics - Fix out-of-bounds dataSet access by replacing reserve+index with push_back/emplace_back - Add proper word boundary detection in mmap-based corpus loader - Split text8 corpus into fixed 1000-word sentence chunks for Skip-Gram training - Add FetchContent CMake integration for fast_float library

+++ TokenEmbeddings/SkipGramModel.h
@@ -11,7 +11,7 @@
 class SkipGramModel {
     public:
 
-    SkipGramModel(int vocabSize, const int &dim, double learningRate, int kNegatives, Tokenizer& tokenizer);
+    SkipGramModel(size_t vocabSize, const int &dim, double learningRate, int kNegatives, Tokenizer& tokenizer);
 
     double trainPairs(int center, int target);
